+++ apps/pxweb2/src/app/components/Pages/Home/Home.module.scss
@@ -1,0 +1,5 @@
+@use '$ui/style-dictionary/dist/scss/fixed-variables.scss' as fixed;
+
+.home {
+  background-color: red;
+}
