+++ apps/pxweb2/src/main.tsx
@@ -1,26 +1,9 @@
-import { StrictMode, Suspense } from 'react';
+import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
+
 import { validateConfig } from './app/util/validate';
 import App from './app/app';
 import './i18n/config';
-import { VariablesProvider } from './app/context/VariablesProvider';
-import { TableDataProvider } from './app/context/TableDataProvider';
-import {
-  createBrowserRouter,
-  Navigate,
-  RouterProvider,
-} from 'react-router-dom';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <Navigate to="/table/tab638" replace={true} />,
-  },
-  {
-    path: '/table/:tableId',
-    element: <App />,
-  },
-]);
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -33,12 +16,6 @@ if (location.href.indexOf('localhost') > -1) {
 
 root.render(
   <StrictMode>
-    <VariablesProvider>
-      <TableDataProvider>
-        <Suspense fallback={<div>Loading...</div>}>
-          <RouterProvider router={router} />
-        </Suspense>
-      </TableDataProvider>
-    </VariablesProvider>
+    <App />
   </StrictMode>
 );
